@@ -12,14 +12,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=2.19.0 <3.0.0'
 
 dependencies:
   url_launcher: ^6.1.7
-  google_fonts: ^3.0.1
+  google_fonts: ^4.0.1
   analyzer: ^5.2.0
   intl: ^0.17.0
-  nylo_framework: ^4.0.0
+  nylo_framework: ^4.1.0
   pretty_dio_logger: ^1.1.1
   flutter:
     sdk: flutter
