@@ -1,4 +1,4 @@
-import '../../config/events.dart';
+import '/config/events.dart';
 import 'package:nylo_framework/nylo_framework.dart';
 
 class EventProvider implements NyProvider {
