@@ -1,5 +1,6 @@
+import 'package:flutter_app/app/models/user.dart';
 import 'package:flutter_app/app/networking/dio/base_api_service.dart';
-import '../app/networking/api_service.dart';
+import '/app/networking/api_service.dart';
 
 /*
 |--------------------------------------------------------------------------
