@@ -1,13 +1,18 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter_app/resources/widgets/app_loader_widget.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_app/resources/widgets/loader_widget.dart';
+import 'package:flutter_app/resources/widgets/logo_widget.dart';
 
 /*
 |--------------------------------------------------------------------------
 | Design
 | Contains widgets used in the Nylo framework.
 |
-| Learn more: https://nylo.dev/docs/4.x/themes
+| Learn more: https://nylo.dev/docs/4.x/themes#design
 |--------------------------------------------------------------------------
 */
 
-Widget loader = AppLoader();
+Widget logo = Logo();
+// resources/widgets/logo_widget.dart
+
+Widget loader = Loader();
+// resources/widgets/loader_widget.dart
