@@ -1,5 +1,5 @@
-import '../app/events/login_event.dart';
-import '../app/events/logout_event.dart';
+import '/app/events/login_event.dart';
+import '/app/events/logout_event.dart';
 import 'package:nylo_framework/nylo_framework.dart';
 
 /*
